+++ CHANGELOG.md
@@ -22,6 +22,9 @@
 ## mini.completion
 
 - FEATURE: respect `isIncomplete` in LSP completion response and immediately force new completion request on the next key press.
+- FEATURE: both info and signature help windows now use tree-sitter highlighting:
+    - Info window uses "markdown" parser (works best on Neovim>=0.10 as its parser is built-in). Special markdown characters are concealed (i.e. hidden) which might result into seemingly unnecessary whitespace as dimensions are computed not accounting for that.
+    - Signature help uses same parser as in current filetype.
 
 ## mini.doc